Add -Zno-codegen hollow rlib pipelining behind incompatible flag
Introduce `incompatible_use_unstable_no_codegen_for_pipelining` to switch
pipelined compilation from the process-wrapper `--rustc-quit-on-rmeta`
approach to Buck2-style `-Zno-codegen` hollow rlibs. Defaults to off; the
existing rmeta-interception path remains the default behavior.

When enabled:
- Metadata action runs `rustc -Zno-codegen --emit=link=<path>` to produce
a hollow rlib (`_meta.rlib` in a `_meta/` subdir of the full rlib).
- Full action emits only `--emit=link` (no metadata, no dep-info).
- `RUSTC_BOOTSTRAP=1` is set on both actions so SVH matches.
- Metadata-consuming actions get an extra `-Ldependency` pointing at the
`_meta/` subdir for transitive resolution.

Tracked in https://github.com/bazelbuild/rules_rust/issues/3977.
