+++ rust/private/rust.bzl
@@ -50,6 +50,7 @@ load(
     "generate_output_diagnostics",
     "get_edition",
     "get_import_macro_deps",
+    "metadata_output_path",
     "transform_deps",
     "transform_sources",
 )
@@ -206,7 +207,7 @@ def _rust_library_common(ctx, crate_type):
         disable_pipelining = getattr(ctx.attr, "disable_pipelining", False),
     ):
         rust_metadata = ctx.actions.declare_file(
-            paths.replace_extension(rust_lib_name, ".rmeta"),
+            metadata_output_path(toolchain, rust_lib_name),
             sibling = rust_lib,
         )
         rustc_rmeta_output = generate_output_diagnostics(ctx, rust_metadata)
@@ -279,7 +280,7 @@ def _rust_binary_impl(ctx):
     rustc_rmeta_output = None
     if can_build_metadata(toolchain, ctx, ctx.attr.crate_type):
         rust_metadata = ctx.actions.declare_file(
-            paths.replace_extension("lib" + crate_name, ".rmeta"),
+            metadata_output_path(toolchain, "lib" + crate_name),
             sibling = output,
         )
         rustc_rmeta_output = generate_output_diagnostics(ctx, rust_metadata)
@@ -381,7 +382,7 @@ def _rust_test_impl(ctx):
         rustc_rmeta_output = None
         if can_build_metadata(toolchain, ctx, crate_type):
             rust_metadata = ctx.actions.declare_file(
-                paths.replace_extension("lib" + crate_name, ".rmeta"),
+                metadata_output_path(toolchain, "lib" + crate_name),
                 sibling = output,
             )
             rustc_rmeta_output = generate_output_diagnostics(ctx, rust_metadata)
@@ -451,7 +452,7 @@ def _rust_test_impl(ctx):
         rustc_rmeta_output = None
         if can_build_metadata(toolchain, ctx, crate_type):
             rust_metadata = ctx.actions.declare_file(
-                paths.replace_extension("lib" + crate_name, ".rmeta"),
+                metadata_output_path(toolchain, "lib" + crate_name),
                 sibling = output,
             )
             rustc_rmeta_output = generate_output_diagnostics(ctx, rust_metadata)
