+++ rust/private/unpretty.bzl
@@ -202,7 +202,7 @@ def _rust_unpretty_aspect_impl(target, ctx):
             out_dir = out_dir,
             build_env_files = build_env_files,
             build_flags_files = build_flags_files,
-            emit = ["dep-info", "metadata"],
+            emit = ["metadata"],
             skip_expanding_rustc_env = True,
         )
