+++ rust/settings/BUILD.bazel
@@ -28,6 +28,7 @@ load(
     "incompatible_change_clippy_error_format",
     "incompatible_do_not_include_data_in_compile_data",
     "incompatible_do_not_include_transitive_data_in_compile_inputs",
+    "incompatible_use_unstable_no_codegen_for_pipelining",
     "lto",
     "no_std",
     "pipelined_compilation",
@@ -113,6 +114,8 @@ incompatible_change_clippy_error_format()
 incompatible_do_not_include_data_in_compile_data()
 
 incompatible_do_not_include_transitive_data_in_compile_inputs()
+
+incompatible_use_unstable_no_codegen_for_pipelining()
 
 lto()
