+++ test/pipelining_bootstrap_gate/unstable.rs
@@ -1,0 +1,3 @@
+#![feature(trait_alias)]
+
+pub trait Alias = Clone + Send;
