+++ test/unit/metadata_output_groups/metadata_output_groups.bzl
@@ -3,6 +3,13 @@
 load("@bazel_skylib//lib:unittest.bzl", "analysistest", "asserts")
 load("//rust:defs.bzl", "rust_binary", "rust_library", "rust_proc_macro", "rust_test")
 
+_TARGETS = [
+    struct(name = "bin", rule = rust_binary, srcs = ["bin.rs"]),
+    struct(name = "lib", rule = rust_library, srcs = ["lib.rs"]),
+    struct(name = "macro", rule = rust_proc_macro, srcs = ["macro.rs"]),
+    struct(name = "unit", rule = rust_test, srcs = ["unit.rs"]),
+]
+
 def _metadata_output_groups_present_test_impl(ctx):
     env = analysistest.begin(ctx)
     tut = analysistest.target_under_test(env)
@@ -12,13 +19,17 @@ def _metadata_output_groups_present_test_impl(ctx):
     rustc_rmeta_output = output_groups.rustc_rmeta_output.to_list()
 
     asserts.equals(env, 1, len(build_metadata), "Expected 1 build_metadata file")
+    basename = build_metadata[0].basename
+    expected_suffix = "_meta.rlib" if ctx.attr.expect_new_path else ".rmeta"
     asserts.true(
-        build_metadata[0].basename.endswith(".rmeta"),
-        "Expected %s to end with .rmeta" % build_metadata[0],
+        env,
+        basename.endswith(expected_suffix),
+        "Expected %s to end with %s" % (build_metadata[0], expected_suffix),
     )
 
     asserts.equals(env, 1, len(rustc_rmeta_output), "Expected 1 rustc_rmeta_output file")
     asserts.true(
+        env,
         rustc_rmeta_output[0].basename.endswith(".rustc-output"),
         "Expected %s to end with .rustc-output" % rustc_rmeta_output[0],
     )
@@ -35,11 +46,23 @@ def _metadata_output_groups_missing_test_impl(ctx):
 
     return analysistest.end(env)
 
-metadata_output_groups_present_test = analysistest.make(
+metadata_output_groups_present_legacy_test = analysistest.make(
     _metadata_output_groups_present_test_impl,
+    attrs = {"expect_new_path": attr.bool(default = False)},
     config_settings = {
         str(Label("//rust/settings:always_enable_metadata_output_groups")): True,
         str(Label("//rust/settings:rustc_output_diagnostics")): True,
+        str(Label("//rust/settings:incompatible_use_unstable_no_codegen_for_pipelining")): False,
+    },
+)
+
+metadata_output_groups_present_new_test = analysistest.make(
+    _metadata_output_groups_present_test_impl,
+    attrs = {"expect_new_path": attr.bool(default = True)},
+    config_settings = {
+        str(Label("//rust/settings:always_enable_metadata_output_groups")): True,
+        str(Label("//rust/settings:rustc_output_diagnostics")): True,
+        str(Label("//rust/settings:incompatible_use_unstable_no_codegen_for_pipelining")): True,
     },
 )
 
@@ -47,63 +70,24 @@ metadata_output_groups_missing_test = analysistest.make(
     _metadata_output_groups_missing_test_impl,
 )
 
-def _output_groups_test(*, always_enable):
-    suffix = "_with_metadata" if always_enable else "_without_metadata"
+def _output_groups_test(*, always_enable, suffix, present_test = None):
+    test = present_test if always_enable else metadata_output_groups_missing_test
 
-    rust_binary(
-        name = "bin" + suffix,
-        srcs = ["bin.rs"],
-        edition = "2021",
-    )
+    for target in _TARGETS:
+        target.rule(
+            name = target.name + suffix,
+            srcs = target.srcs,
+            edition = "2021",
+        )
 
-    rust_library(
-        name = "lib" + suffix,
-        srcs = ["lib.rs"],
-        edition = "2021",
-    )
-
-    rust_proc_macro(
-        name = "macro" + suffix,
-        srcs = ["macro.rs"],
-        edition = "2021",
-    )
-
-    rust_test(
-        name = "unit" + suffix,
-        srcs = ["unit.rs"],
-        edition = "2021",
-    )
-
-    if always_enable:
-        test = metadata_output_groups_present_test
-    else:
-        test = metadata_output_groups_missing_test
-
-    test(
-        name = "bin_test" + suffix,
-        target_under_test = ":bin" + suffix,
-    )
-
-    test(
-        name = "lib_test" + suffix,
-        target_under_test = ":lib" + suffix,
-    )
-
-    test(
-        name = "macro_test" + suffix,
-        target_under_test = ":macro" + suffix,
-    )
-
-    test(
-        name = "unit_test" + suffix,
-        target_under_test = ":unit" + suffix,
-    )
+        test(
+            name = target.name + "_test" + suffix,
+            target_under_test = ":" + target.name + suffix,
+        )
 
     return [
-        ":bin_test" + suffix,
-        ":lib_test" + suffix,
-        ":macro_test" + suffix,
-        ":unit_test" + suffix,
+        ":" + target.name + "_test" + suffix
+        for target in _TARGETS
     ]
 
 def metadata_output_groups_test_suite(name):
@@ -113,8 +97,20 @@ def metadata_output_groups_test_suite(name):
         name: Name of the macro.
     """
     tests = []
-    tests.extend(_output_groups_test(always_enable = True))
-    tests.extend(_output_groups_test(always_enable = False))
+    tests.extend(_output_groups_test(
+        always_enable = True,
+        suffix = "_with_metadata_legacy",
+        present_test = metadata_output_groups_present_legacy_test,
+    ))
+    tests.extend(_output_groups_test(
+        always_enable = True,
+        suffix = "_with_metadata_new",
+        present_test = metadata_output_groups_present_new_test,
+    ))
+    tests.extend(_output_groups_test(
+        always_enable = False,
+        suffix = "_without_metadata",
+    ))
 
     native.test_suite(
         name = name,
